[Builtin_cd] hide builtin_cd_rule_8
