+++ inc/builtins.h
@@ -41,7 +41,6 @@ int				builtin_setenv(int argc, const char * const *argv);
 int				builtin_unsetenv(int argc, const char * const *argv);
 
 int				builtin_cd(int argc, const char * const *argv);
-int				builtin_cd_rule_8(t_string *curpath);
 const char		*get_next_component(t_string *component, const char *path);
 void			get_base_path(t_string *curpath, bool *must_print_pwd,
 								const char *dir);
